test: organize coverage by behavior and feature
