+++ tests/renderer/prepare-columns.test.ts
@@ -1,8 +1,8 @@
 import { expect, test } from "bun:test";
-import type { CellInfo, ColumnDef } from "../src/columns/types";
-import { TaskId } from "../src/task-model";
-import { resolveColumns } from "../src/renderer/column-layout";
-import { makeRows, makeTaskSnapshot } from "./helpers/renderer";
+import type { CellInfo, ColumnDef } from "../../src/columns/types";
+import { TaskId } from "../../src/task-model";
+import { resolveColumns } from "../../src/renderer/column-layout";
+import { makeRows, makeTaskSnapshot } from "../helpers/renderer";
 
 test("shared preparation runs once per position and stays bound to each renderer and sizing hint", () => {
   const calls: number[] = [];
